--- a/hyperspace/testsuite/tests/parachain_parachain.rs
+++ b/hyperspace/testsuite/tests/parachain_parachain.rs
@@ -25,6 +25,7 @@ use hyperspace_testsuite::{
 	ibc_messaging_packet_timestamp_timeout_with_connection_delay,
 	ibc_messaging_with_connection_delay, misbehaviour::ibc_messaging_submit_misbehaviour,
 };
+use std::time::Duration;
 
 #[derive(Debug, Clone)]
 pub struct Args {
@@ -130,31 +131,78 @@ async fn setup_clients() -> (ParachainClient<DefaultConfig>, ParachainClient<Def
 
 #[tokio::test]
 async fn parachain_to_parachain_ibc_messaging_full_integration_test() {
+	use std::time::SystemTime;
 	logging::setup_logging();
+	use hyperspace_testsuite::setup_connection_and_channel;
+	use ibc::core::ics24_host::identifier::PortId;
 	let (mut chain_a, mut chain_b) = setup_clients().await;
-	// Run tests sequentially
+	let mut chain_aa = chain_a.clone();
+	let mut chain_bb = chain_b.clone();
+	//set up connection only once!!!
+	let (handle, channel_a, channel_b, connection_id_a, connection_id_b) =
+		setup_connection_and_channel(&mut chain_a, &mut chain_b, Duration::from_secs(60 * 2)).await;
+	handle.abort();
+
+	// Set connections and channel whitelist
+	chain_a.set_connection_id(connection_id_a);
+	chain_b.set_connection_id(connection_id_b);
+
+	chain_a.set_channel_whitelist(vec![(channel_a, PortId::transfer())]);
+	chain_b.set_channel_whitelist(vec![(channel_b, PortId::transfer())]);
 
 	let asset_id = 1;
 
-	// let join_set = JoinSet::new();
+	let mut join_set = tokio::task::JoinSet::new();
 
 	// no timeouts + connection delay
-	ibc_messaging_with_connection_delay(&mut chain_a, &mut chain_b, asset_id, asset_id).await;
+	let mut c1 = chain_a.clone();
+	let mut c2 = chain_b.clone();
+	join_set.spawn(async move {
+		ibc_messaging_with_connection_delay(
+			&mut c1, &mut c2, asset_id, asset_id, channel_a, channel_b,
+		)
+		.await;
+	});
 
 	// timeouts + connection delay
-	ibc_messaging_packet_height_timeout_with_connection_delay(&mut chain_a, &mut chain_b, asset_id)
+	let mut c1 = chain_a.clone();
+	let mut c2 = chain_b.clone();
+	join_set.spawn(async move {
+		ibc_messaging_packet_height_timeout_with_connection_delay(
+			&mut c1, &mut c2, asset_id, channel_a, channel_b,
+		)
 		.await;
+	});
 
-	ibc_messaging_packet_timestamp_timeout_with_connection_delay(
-		&mut chain_a,
-		&mut chain_b,
-		asset_id,
-	)
-	.await;
+	let mut c1 = chain_a.clone();
+	let mut c2 = chain_b.clone();
+	join_set.spawn(async move {
+		ibc_messaging_packet_timestamp_timeout_with_connection_delay(
+			&mut c1, &mut c2, asset_id, channel_a, channel_b,
+		)
+		.await;
+	});
+
+	log::info!(target: "hyperspace", "🚀🚀 Waiting for the 3 features.");
+	while let Some(res) = join_set.join_next().await {
+		res.unwrap();
+	}
 
 	// channel closing semantics
-	ibc_messaging_packet_timeout_on_channel_close(&mut chain_a, &mut chain_b, asset_id).await;
-	ibc_channel_close(&mut chain_a, &mut chain_b).await;
+	let mut join_set = tokio::task::JoinSet::new();
+	let mut c1 = chain_a.clone();
+	let mut c2 = chain_b.clone();
+	join_set.spawn(async move {
+		ibc_messaging_packet_timeout_on_channel_close(&mut c1, &mut c2, asset_id, channel_a).await;
+	});
+	join_set.spawn(async move {
+		ibc_channel_close(&mut chain_aa, &mut chain_bb).await;
+	});
+
+	log::info!(target: "hyperspace", "🚀🚀 Waiting for the 2 features.");
+	while let Some(res) = join_set.join_next().await {
+		res.unwrap();
+	}
 
 	// Test sync abilities, run this before misbehaviour test
 	client_synchronization_test(&mut chain_a, &mut chain_b).await;